requirements_extended.txt
# Existing requirements
sentence-transformers
faiss-cpu
numpy

# Additional requirements for OCR and video processing
opencv-python
pytesseract
Pillow
pathlib

# Optional: For better OCR preprocessing
scikit-image